chore(warehouse): correct migration name ,managing nil values in jobs

+++ internal/db/queries.sql
@@ -349,7 +349,7 @@ WITH t AS (
 		@steps::JSONB,
 		@checkrunurl::TEXT,
 		@labels::JSONB,
-		@runnerid::INTEGER,
+		@runnerid::BIGINT,
 		@runnername::TEXT,
 		@runnergroupid::BIGINT,
 		@runnergroupname::TEXT)
